# backend/docs/swagger.yaml
basePath: /
...
host: localhost:8000
info:
  contact:
    email: support@lagertool.com
    name: API Support
  description: Backend API for inventory management system tracking items, locations,
    and loans
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Lagertool Inventory API
  version: "1.0"
paths:
  /calendar/{id}:
    get:
      description: Downloads an iCalendar (.ics) file for a specific loan containing
        a reminder to return the item
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/calendar
      responses:
        "200":
          description: ICS file content
          schema:
            type: string
        "400":
          description: Invalid loan ID
          schema:
            type: string
      summary: Download calendar file for a specific loan
      tags:
      - calendar
...
schemes:
- http
swagger: "2.0"